Move Configuration tests to Configuration module
