+++ configuration/src/test/java/com/shieldblaze/expressgateway/configuration/eventloop/EventLoopConfigurationBuilderTest.java
@@ -1,0 +1,48 @@
+/*
+ * This file is part of ShieldBlaze ExpressGateway. [www.shieldblaze.com]
+ * Copyright (c) 2020 ShieldBlaze
+ *
+ * ShieldBlaze ExpressGateway is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * ShieldBlaze ExpressGateway is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with ShieldBlaze ExpressGateway.  If not, see <https://www.gnu.org/licenses/>.
+ */
+package com.shieldblaze.expressgateway.configuration.eventloop;
+
+import com.shieldblaze.expressgateway.configuration.eventloop.EventLoopConfigurationBuilder;
+import org.junit.jupiter.api.Test;
+
+import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
+import static org.junit.jupiter.api.Assertions.assertThrows;
+
+final class EventLoopConfigurationBuilderTest {
+
+    @Test
+    void build() {
+        assertThrows(IllegalArgumentException.class, () -> EventLoopConfigurationBuilder.newBuilder()
+                .withParentWorkers(0)
+                .build());
+
+        assertThrows(IllegalArgumentException.class, () -> EventLoopConfigurationBuilder.newBuilder()
+                .withParentWorkers(1)
+                .build());
+
+        assertThrows(IllegalArgumentException.class, () -> EventLoopConfigurationBuilder.newBuilder()
+                .withParentWorkers(1)
+                .withChildWorkers(0)
+                .build());
+
+        assertDoesNotThrow(() -> EventLoopConfigurationBuilder.newBuilder()
+                .withParentWorkers(1)
+                .withChildWorkers(1)
+                .build());
+    }
+}
